- svn:eol-style = native fixes - really annoying cast in font.cpp (they don't need to be floats!) - move #ifdef OPENGL inside header guards to enable optimization (some compilers are exactly half-stupid)
SVN-Revision: 5774

+++ src/video/font.cpp
@@ -232,8 +232,8 @@ Font::loadFontSurface(
 
   for( unsigned int i = 0; i < chars.size(); i++) {
     for(UTF8Iterator chr(chars[i]); !chr.done(); ++chr) {
-      float y = row * char_height;
-      float x = col * char_width;
+      int y = row * char_height;
+      int x = col * char_width;
       if( ++col == wrap ) { col=0; row++; }
       if( *chr == 0x0020 && glyphs[0x20].surface_idx != -1) continue;
